Revert "Change register responce. Refactoring"
