+++ src/main/java/com/epam/dto/EcoUserDTO.java
@@ -1,12 +1,19 @@
 package com.epam.dto;
 
+import com.epam.model.EcoService;
 import com.epam.security.EcoUserRole;
 import com.fasterxml.jackson.annotation.JsonAlias;
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.annotation.JsonView;
 import lombok.Getter;
 import lombok.Setter;
 
+import javax.persistence.CascadeType;
+import javax.persistence.FetchType;
+import javax.persistence.OneToMany;
 import javax.validation.constraints.*;
+import java.util.List;
 
 @Setter
 @Getter
@@ -22,12 +29,12 @@ public class EcoUserDTO {
     @Size(min = 9, max = 24, message = "PasswordSizeError")
     private String password;
 
-    @NotBlank(message = "EmailEmptyError")
+    @NotBlank
     @Email(regexp = "^[a-zA-Z\\d.]+@[a-zA-Z\\d.]+$", message = "EmailError")
+    @NotNull(message = "EmailError")
     private String email;
 
     @JsonAlias("role")
     @JsonProperty("role")
-    @NotNull(message = "RoleError")
-    private EcoUserRole userRole;
+    private EcoUserRole userRole=EcoUserRole.ANONYMOUS;
 }
