+++ src/main/java/com/epam/web/JwtAuthenticationController.java
@@ -1,12 +1,6 @@
 package com.epam.web;
 
 
-import com.epam.dto.EcoUserDTO;
-import com.epam.model.JwtRequest;
-import com.epam.model.JwtResponse;
-import com.epam.service.JwtEcoUserDetailsService;
-import com.epam.service.JwtEcoUserService;
-import lombok.Getter;
 import com.epam.config.JwtTokenUtil;
 import com.epam.dto.JwtAuthRequest;
 import com.epam.dto.JwtControllersResponseMessage;
@@ -17,72 +11,61 @@ import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.BadCredentialsException;
+import org.springframework.security.authentication.LockedException;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.validation.BindException;
-import org.springframework.validation.ObjectError;
 import org.springframework.web.bind.annotation.*;
-import javax.validation.Valid;
-import java.util.Collections;
-import java.util.Set;
-import java.util.stream.Collectors;
-
 
 @RestController
 @RequestMapping("/api/auth/")
 @RequiredArgsConstructor
 public class JwtAuthenticationController {
 
-  private final JwtEcoUserService jwtEcoUserService;
+	private final AuthenticationManager authenticationManager;
 
-  @Getter
-  private static final class ResponseMessage {
-    private final Set<String> message;
+	private final JwtTokenUtil jwtTokenUtil;
 
-    ResponseMessage(BindException e) {
-      this.message =
-          e.getAllErrors().stream().map(ObjectError::getDefaultMessage).collect(Collectors.toSet());
-    }
+	private final JwtEcoUserDetailsService jwtEcoUserDetailsService;
 
-    ResponseMessage(Exception e) {
-      this(e.getMessage());
-    }
+	@RequestMapping(value = "/authenticate", method = RequestMethod.POST)
+	public ResponseEntity<?> createAuthenticationToken(@RequestBody final JwtAuthRequest authenticationRequest) throws Exception {
 
-    ResponseMessage(String message) {
-      this.message = Collections.singleton(message);
-    }
-  }
+		authenticate(authenticationRequest.getUsername(), authenticationRequest.getPassword());
 
-  @RequestMapping(value = "/authenticate", method = RequestMethod.POST)
-  public ResponseEntity<?> createAuthenticationToken(@RequestBody final JwtRequest authenticationRequest) {
-    return ResponseEntity.ok(
-        new JwtResponse(
-            jwtEcoUserService.getToken(
-                authenticationRequest.getUsername(), authenticationRequest.getPassword())));
-  }
+		final UserDetails userDetails = jwtEcoUserDetailsService
+				.loadUserByUsername(authenticationRequest.getUsername());
 
-  @RequestMapping(value = "/register", method = RequestMethod.POST)
-  public ResponseEntity<?> saveUser(@Valid @RequestBody final EcoUserDTO user) {
-    jwtEcoUserService.save(user);
-    return ResponseEntity.ok(
-            new JwtResponse(
-                    jwtEcoUserService.getToken(
-                            user.getUsername(), user.getPassword())));
-  }
+		final String token = jwtTokenUtil.generateToken(userDetails);
 
-  @ExceptionHandler(AuthenticationException.class)
-  public ResponseEntity<?> handleAuthenticationExceptionException(final AuthenticationException e) {
-    return new ResponseEntity<>(new ResponseMessage(e), HttpStatus.UNAUTHORIZED);
-  }
+		return ResponseEntity.ok(new JwtTockenResponse(token));
+	}
 
-  @ExceptionHandler(BindException.class)
-  public ResponseEntity<?> handleMethodArgumentNotValidException(final BindException e) {
-    return new ResponseEntity<>(new ResponseMessage(e), HttpStatus.UNAUTHORIZED);
-  }
+	private void authenticate(final String username, final String password) throws Exception {
+		try {
+			authenticationManager.authenticate(new UsernamePasswordAuthenticationToken(username, password));
+		} catch (LockedException e) {
+			throw new LockedException("USER_LOCKED", e);
+		} catch (BadCredentialsException e) {
+			throw new BadCredentialsException("INVALID_CREDENTIALS", e);
+		}
+	}
 
-  @ExceptionHandler(Exception.class)
-  public ResponseEntity<?> handleException(final Exception e) {
-    final Log logger = LogFactory.getLog(this.getClass());
-    logger.error(e);
-    return new ResponseEntity<>(new ResponseMessage("COMMON_ERROR"), HttpStatus.UNAUTHORIZED);
-  }
+	@ExceptionHandler(AuthenticationException.class)
+	public ResponseEntity<?> handleAuthenticationExceptionException(final AuthenticationException e) {
+		return new ResponseEntity<>(new JwtControllersResponseMessage(e), HttpStatus.UNAUTHORIZED);
+	}
+
+	@ExceptionHandler(BindException.class)
+	public ResponseEntity<?> handleMethodArgumentNotValidException(final BindException e) {
+		return new ResponseEntity<>(new JwtControllersResponseMessage(e), HttpStatus.UNAUTHORIZED);
+	}
+
+	@ExceptionHandler(Exception.class)
+	public ResponseEntity<?> handleException(final Exception e) {
+		return new ResponseEntity<>(new JwtControllersResponseMessage("COMMON_ERROR"), HttpStatus.UNAUTHORIZED);
+	}
 }
